# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "业界优秀的通用管理框架基座（非业务系统）：RBAC 权限、菜单、部门、字典、操作日志、登录日志、数据范围、系统配置。\n清洁架构 handler → service → repository → model，复用泛型 Repository 基类。",
        "title": "GVA 通用管理框架基座 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api",
    "paths": {
...
        "/system/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sys-config"
                ],
                "summary": "系统配置分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "key/name 关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sys-config"
                ],
                "summary": "创建系统配置",
                "parameters": [
                    {
                        "description": "配置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.sysConfigUpsertReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    },
                    "409": {
                        "description": "key 已存在",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ProblemDetail"
                        }
                    }
                }
            }
        },
        "/system/config/key/{key}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sys-config"
                ],
                "summary": "按 key 取配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置 key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ProblemDetail"
                        }
                    }
                }
            }
        },
        "/system/config/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sys-config"
                ],
                "summary": "系统配置详情（按 ID）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ProblemDetail"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sys-config"
                ],
                "summary": "更新系统配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.sysConfigUpsertReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ProblemDetail"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sys-config"
                ],
                "summary": "删除系统配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ApiResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gva_internal_pkg_response.ProblemDetail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gva_internal_pkg_response.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "gva_internal_pkg_response.ProblemDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "应用层错误码（机器可读）",
                    "type": "string"
                },
                "detail": {
                    "description": "具体说明",
                    "type": "string"
                },
                "errors": {
                    "description": "字段级错误（表单校验）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "description": "HTTP 状态码",
                    "type": "integer"
                },
                "title": {
                    "description": "简短摘要",
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                },
                "type": {
                    "description": "问题类型 URI",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.sysConfigUpsertReq": {
            "type": "object",
            "required": [
                "configKey"
            ],
            "properties": {
                "configKey": {
                    "type": "string",
                    "maxLength": 128
                },
                "configName": {
                    "type": "string",
                    "maxLength": 128
                },
                "configValue": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {accessToken}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}